move ReportContext to loadflow.server.computation
