+++ src/main/java/org/gridsuite/loadflow/server/LoadFlowController.java
@@ -16,7 +16,7 @@ import org.gridsuite.loadflow.server.dto.LoadFlowResult;
 import org.gridsuite.loadflow.server.dto.LoadFlowStatus;
 import org.gridsuite.loadflow.server.service.LoadFlowRunContext;
 import org.gridsuite.loadflow.server.service.LoadFlowService;
-import org.gridsuite.loadflow.server.utils.ReportContext;
+import org.gridsuite.loadflow.server.computation.utils.ReportContext;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
