+++ src/main/java/org/gridsuite/loadflow/server/computation/utils/ReportContext.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package org.gridsuite.loadflow.server.utils;
+package org.gridsuite.loadflow.server.computation.utils;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
