+++ src/main/java/org/gridsuite/loadflow/server/service/LoadFlowRunContext.java
@@ -14,7 +14,7 @@ import lombok.Builder;
 import lombok.Getter;
 import org.gridsuite.loadflow.server.dto.parameters.LoadFlowParametersValues;
 import org.gridsuite.loadflow.server.computation.service.AbstractComputationRunContext;
-import org.gridsuite.loadflow.server.utils.ReportContext;
+import org.gridsuite.loadflow.server.computation.utils.ReportContext;
 
 import java.util.UUID;
